requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# AI and LangChain
#anthropic==0.18.1
langchain==0.1.6
#langchain-anthropic==0.1.1
langchain-community==0.0.19
google-generativeai==0.3.2

# Vector Store (for future Phase 2)
chromadb==0.4.22

# Environment and Config
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# CORS and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
httpx==0.26.0
aiofiles==23.2.1